+++ CHANGELOG.md
@@ -5,6 +5,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2026-06-03
+
+### Added
+
+- New `StepConstant`: `PreTrigger`, `PostTrigger`, `Trigger`, `CurrentCalibre`.
+- New `Instrument`: `Sync`, holding the trigger command. Has a `trigger`
+  property that computes the value of the trigger.
+
+### Changed
+
+- `HeaderConstant` is now `StepConstant`
+- Frequency should be given in `TOML` via a `FrequencySetpoint`, even though
+  it can still be given as a `MultipactorTest` argument.
+- Old `PowerSetpoint` inheriting from `VirtualInstrument` are removed. They are
+  replaced by equivalent `PowerSetpoint` inheriting from `StepConstant`.
+
+### Removed
+
+- `Frequency` implementation. This is now an alias to `FrequencySetpoint`.
+
 ## [2.0.0] - 2026-06-03
 
 ### Added